feat: modify some service with preload, and add owner to team with service adapted

+++ app/services/team_service.ts
@@ -1,5 +1,6 @@
 import Team from '#models/team'
 import User from '#models/user'
+import type { ModelQueryBuilderContract } from '@adonisjs/lucid/types/model'
 
 /**
  * Service class for managing teams
@@ -19,7 +20,20 @@ export default class TeamService {
    * @returns {Promise<Team>} - A promise that resolves with the team
    */
   public static async getTeamById(teamId: number): Promise<Team> {
-    return await Team.query().where('id', teamId).preload('users').firstOrFail()
+    return await Team.query().where('id', teamId).preload('users').preload('projects').firstOrFail()
+  }
+
+  /**
+   * Récupère toutes les équipes auxquelles appartient un utilisateur.
+   * @param {number} userId - L'ID de l'utilisateur
+   * @returns {Promise<Team[]>} - Une promesse qui résout avec un tableau d'équipes
+   */
+  public static async getTeamsByUserId(userId: number): Promise<Team[]> {
+    return Team.query()
+      .preload('users')
+      .whereHas('users', (query: ModelQueryBuilderContract<typeof User>) => {
+        query.where('id', userId)
+      })
   }
 
   /**
@@ -27,12 +41,14 @@ export default class TeamService {
    * @param {string} data - The data to create the team
    * @param {string} data.name - The name of the team
    * @param {string} data.description - The description of the team
+   * @param {number} data.owner_id - The ID of the team owner
    * @returns {Promise<Team>} - A promise that resolves with the created team
    */
-  public static async createTeam(data: { name: string; description?: string }): Promise<Team> {
+  public static async createTeam(data: { name: string; description?: string; owner_id: number }): Promise<Team> {
     const team: Team = new Team()
     team.name = data.name
     team.description = data.description
+    team.owner_id = data.owner_id
     await team.save()
     return team
   }
@@ -59,9 +75,13 @@ export default class TeamService {
   /**
    * Delete a team
    * @param {number} teamId - The ID of the team
+   * @param {number} owner_id - The ID of the team owner
    */
-  public static async deleteTeam(teamId: number): Promise<void> {
+  public static async deleteTeam(teamId: number, owner_id: number): Promise<void> {
     const team: Team = await Team.findOrFail(teamId)
+    if (team.owner_id !== owner_id) {
+      throw new Error('You are not authorized to delete this team')
+    }
     await team.delete()
   }
